bugfix: debug proxy doesn't return proper cert.
we must use Let's Encrypt full chain of certificate, otherwise the
Android client rejects the certificate.
the greenlock's server uses the proper certificate (and thus the
productioin server works OK), but for debug server (on port 8090) we
load the cert manually - and must use the fullchain.

NOTE: there's a "known issue" for the debugging proxy, and that it
doesn't reload the certificate every ~2 months. greenlock will reload it
for the production server, but the debug server might get into outdated
certifcate.
restarting the service (`sudo systemctl restart bizpoc-proxy
`) will to the job, as would `sudo pkill -f node`