Initial support for SafetyNet API in the app
Created SafetyNet helper that performs the request and handles the
initial prsing of the JWT token.
During provisioning, the app will proceed to "getEncryptedCredentials"
only affter obtaining and locally checking the SafetyNet attestation.
After provisioning, the app will initiate the attestation process once
started but make first batch of requests by reusing the last known
attestation, so we don't wait for it to complete.