Support resuming actication in case verification stalled; v. 1.1.2
In case the application did receive the credentials from the server, it
should not lose them in any case. Also, it would be dangerous to let the
user into the app on the next run without running verification.

So, now the application recognizes this state and goes to the activation
screen, but starts from task-3 (validating credentials).