Add remote nonce generation for SafetyNet
Add new API, "/newnonce/", that returns a string that should be used as
nonce for SafetyNet.

On "/key/" request, instead of it, the HMAC.getBytes() is used for
nonce.

There is an option to verify the timestama, byt in currrent
configuration it is not used by proxy - get /wallet/ does not require a
new attestation.
Provisioning server will be required to provide a fresh attestation
(fresh is defined as 1 day now, con be changed)