BPOC-74: App crashes on returning from background
The following 4 bariables: "Balances", "Exchange Rates", "Guardians" and
"BitGo wallet" ended up being passed around and used as members of every
fragment. Once fragment got re-created by Andriod, the app crashed with
NPE.
There is no functional reason to not keep them as static variables
(initialized by First Fragment so far).
Needs better refactoring.

On the "Confirm" and "Details" fragments, there are context-specific
fields ("Transaction", "Pending Approval" and "Send Request") that
cannot be saved in a static context. Instead, they are now put in the
'savedInstanceState' bundle as Serializable object, as it used to be done.